+++ src/test/scala/scodec/CodecTest.scala
@@ -23,4 +23,26 @@ class CodecTest extends CodecSuite {
   test("as on single codec") {
     roundtripAll(uint8.hlist.as[Bar], Seq(Bar(0), Bar(1), Bar(255)))
   }
+
+  test("unit combinator") {
+    import scalaz.std.AllInstances._
+    val codec = uint8.unitM
+    codec.encode(()) shouldBe \/.right(BitVector(0))
+    codec.decode(BitVector(1)) shouldBe \/.right((BitVector.empty, ()))
+    codec.decode(BitVector.empty) shouldBe 'left
+    uint8.unit(255).encode(()) shouldBe \/.right(BitVector(0xff))
+  }
+
+  test("<~") {
+    val codec = uint8 <~ uint8.unit(0)
+    codec.encode(0xff) shouldBe \/.right(hex"ff00".bits)
+  }
+
+  test("literals used as unit codecs") {
+    import scodec.codecs.literals._
+    (1 ~> uint8).encode(2) shouldBe \/.right(hex"0102".bits)
+    (1.toByte ~> uint8).encode(2) shouldBe \/.right(hex"0102".bits)
+    (hex"11223344" ~> uint8).encode(2) shouldBe \/.right(hex"1122334402".bits)
+    (hex"11223344".bits ~> uint8).encode(2) shouldBe \/.right(hex"1122334402".bits)
+  }
 }